Refactor run command into new management module
The new `run` defined a couple of supporting classes and was a fairly
long function. To keep `./manage.py` readable, this commit factors the
`run` function and supporting code into a new `management` module. This
keeps the `./manage.py` lightweight and easy to read, and better
encapsulates the code specific to the `run` function.
